+++ src/collective/easyform/tests/testDocTests.py
@@ -5,11 +5,11 @@ from plone.testing import layered
 from plone.testing.z2 import Browser
 
 import doctest
+import os
 import re
 import six
 import transaction
 import unittest
-import os
 
 
 optionflags = (
@@ -51,7 +51,7 @@ def get_browser(layer, auth=True):
 
 def get_image_path():
     dir_name = os.path.dirname(os.path.realpath(__file__))
-    return '{0}/PloneLogo.png'.format(dir_name)
+    return "{0}/PloneLogo.png".format(dir_name)
 
 
 def test_suite():
@@ -62,8 +62,10 @@ def test_suite():
                 doctest.DocFileSuite(
                     f,
                     optionflags=optionflags,
-                    globs={"get_browser": get_browser,
-                           "get_image_path": get_image_path},
+                    globs={
+                        "get_browser": get_browser,
+                        "get_image_path": get_image_path,
+                    },
                     checker=Py23DocChecker(),
                 ),
                 layer=FUNCTIONAL_TESTING,