+++ src/collective/easyform/browser/view.py
@@ -323,6 +323,8 @@ class EasyFormForm(AutoExtensibleForm, form.Form):
 
     def header_injection(self):
         tal_expression = self.context.headerInjection
+        if not tal_expression:
+            return ""
         header_to_inject = get_expression(self.context, tal_expression)
         if six.PY2 and isinstance(header_to_inject, six.text_type):
             header_to_inject = header_to_inject.encode("utf-8")